RBM Refactoring
