+++ configs/engine/divae_training.yaml
@@ -5,3 +5,4 @@ n_epochs: 2
 n_gibbs_sampling_steps: 100
 sampling_mode: gibbs_flat
 weight_decay_factor: 0.0001
+momentum_coefficient: 0.5
